+++ monoman.config.ts
@@ -196,6 +196,7 @@ Please refer to [README.md](${githubLink}#readme)\n`
     include: [
       'package.json',
       'packages/*/package.json',
+      'docs/package.json',
       'playground/*/package.json',
     ],
     exclude: ['playground/vue2/package.json'],
@@ -230,13 +231,13 @@ Please refer to [README.md](${githubLink}#readme)\n`
       /shiki/,
       /babel/,
       /esbuild/,
-      /vite(?!-plugin-vue-inspector)/,
+      /vite(?!-plugin-(vue-inspector|inspect))/,
       /unocss/,
       /rolldown/,
     ],
   }),
   ...noDuplicatedPnpmLockfile({
-    deps: ['vue', 'lru-cache', 'minimatch', 'debug'],
+    deps: ['vue', '@vue/compiler-sfc', 'lru-cache', 'minimatch', 'debug'],
     allowMajor: true,
   }),
 ])