+++ packages/api/src/ts/is.ts
@@ -1,6 +1,7 @@
-import { isDeclaration, type Declaration, type TypeScript } from '@babel/types'
+import { isDeclarationType } from '@vue-macros/common'
+import type { Declaration, TypeScript } from '@babel/types'
 
 export type TSDeclaration = TypeScript & Declaration
 export function isTSDeclaration(node: any): node is TSDeclaration {
-  return isDeclaration(node) && node.type.startsWith('TS')
+  return isDeclarationType(node) && node.type.startsWith('TS')
 }