+++ packages/jsx-directive/tests/fixtures/v-if/v-else-if.vue
@@ -1,11 +1,13 @@
+<script setup lang="tsx">
 const { foo = 0 } = defineProps<{
   foo: number
 }>()
 
-export default (
+defineRender(() => (
   <>
     <div v-if={foo === 0}>0</div>
     <div v-else-if={foo === 1}>1</div>
     <div v-else-if={foo === 2}>2</div>
   </>
-)
+))
+</script>