+++ packages/macros/src/index.ts
@@ -98,6 +98,7 @@ const plugin: UnpluginCombineInstance<Options | undefined> =
       resolvePlugin(VueDefineStyleX, framework, options.defineStyleX),
       resolvePlugin(VueExportRender, framework, options.exportRender),
       resolvePlugin(VueExportExpose, framework, options.exportExpose),
+      resolvePlugin(VueJsxDirective, framework, options.jsxDirective),
       resolvePlugin(
         VueReactivityTransform,
         framework,
@@ -133,7 +134,6 @@ const plugin: UnpluginCombineInstance<Options | undefined> =
 
       options.plugins.vue,
       resolvePlugin(VueJsxMacros, framework, options.jsxMacros),
-      resolvePlugin(VueJsxDirective, framework, options.jsxDirective),
       options.plugins.vueJsx,
       resolvePlugin(VueDefineRender, framework, options.defineRender),
       setupComponentPlugins?.[1],